feat: implement error handling component for authentication errors; add UI and routing logic; add dockerfile
