+++ .dockerignore
@@ -1,0 +1,6 @@
+node_modules
+dist
+.next
+.env
+.git
+*.log
