+++ Docker/Dockerfile.frontend
@@ -1,0 +1,22 @@
+FROM node:20-alpine
+
+WORKDIR /usr/src/app
+
+RUN npm install -g pnpm
+
+COPY ./package.json ./package.json
+COPY ./turbo.json ./turbo.json
+COPY ./pnpm-workspace.yaml ./pnpm-workspace.yaml
+COPY ./pnpm-lock.yaml ./pnpm-lock.yaml
+
+COPY ./packages ./packages
+COPY ./apps/web ./apps/web
+
+RUN pnpm install
+RUN pnpm run prisma:generate
+
+RUN pnpm run build
+
+EXPOSE 3000
+
+CMD ["pnpm", "run", "start:fe"]
