+++ Docker/Dockerfile.server
@@ -1,0 +1,23 @@
+FROM node:20-alpine
+
+WORKDIR /usr/src/app
+
+RUN npm install -g pnpm
+
+COPY ./package.json ./package.json
+COPY ./turbo.json ./turbo.json
+COPY ./pnpm-lock.yaml ./pnpm-lock.yaml
+COPY ./pnpm-workspace.yaml ./pnpm-workspace.yaml
+
+
+COPY ./packages ./packages
+COPY ./apps/server ./apps/server
+
+RUN pnpm install
+RUN pnpm run prisma:generate
+
+RUN pnpm run build
+
+EXPOSE 3001
+
+CMD ["pnpm", "run", "start:server"]
