+++ apps/web/app/auth/error/page.tsx
@@ -1,31 +1,13 @@
-'use client';
+"use client";
+import AuthError from "@/components/Error";
+import Loading from "@/components/Loading";
+import { Suspense } from "react";
 
-import { useSearchParams } from 'next/navigation';
-
-export default function AuthError() {
-  const searchParams = useSearchParams();
-  const error = searchParams.get('error');
-
-  return (
-    <div className="flex min-h-screen items-center justify-center bg-gray-50">
-      <div className="w-full max-w-md space-y-8 rounded-lg bg-white p-6 shadow-lg">
-        <div>
-          <h2 className="mt-6 text-center text-3xl font-bold tracking-tight text-red-600">
-            Authentication Error
-          </h2>
-          <p className="mt-2 text-center text-sm text-gray-600">
-            {error || 'An error occurred during authentication'}
-          </p>
-        </div>
-        <div className="mt-8">
-          <a
-            href="/auth/signin"
-            className="group relative flex w-full justify-center rounded-md border border-transparent bg-indigo-600 px-4 py-2 text-sm font-medium text-white hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
-          >
-            Try Again
-          </a>
-        </div>
-      </div>
-    </div>
-  );
-} 
+export default function ErrorPage() {
+  return(
+    <Suspense fallback={<Loading/>}
+    >
+      <AuthError/>
+    </Suspense>
+  )
+}
