--- apps/web/components/providers/ToastProvider.tsx
@@ -1,18 +1,0 @@
-'use client';
-
-import { Toaster } from 'react-hot-toast';
-
-export function ToastProvider() {
-  return (
-    <Toaster
-      position="top-right"
-      toastOptions={{
-        duration: 3000,
-        style: {
-          background: '#333',
-          color: '#fff',
-        },
-      }}
-    />
-  );
-}
