+++ packages/api-client/package.json
@@ -16,7 +16,7 @@
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "type": "module",
+
   "dependencies": {
     "axios": "^1.9.0"
   }
