+++ packages/api-client/src/index.ts
@@ -1,4 +1,4 @@
-export { apiCategory } from "./categories.api.ts";
-export { apiEmail } from "./email.api.ts";
-export { apiAI } from "./ai.api.ts";
+export { apiCategory } from "./categories.api";
+export { apiEmail } from "./email.api";
+export { apiAI } from "./ai.api";
