+++ packages/api-client/tsconfig.json
@@ -1,6 +1,3 @@
 {
     "extends": "@repo/typescript-config/base.json",
-    "compilerOptions": {
-        "allowImportingTsExtensions": true
-    }    
 }
