+++ packages/store/package.json
@@ -16,7 +16,6 @@
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "type": "module",
   "dependencies": {
     "zustand": "^5.0.5"
   }
